montando jar

+++ src/main/resources/db/migration/V6__create-table-consulta.sql
@@ -2,7 +2,7 @@ create table consultas(
 
     id bigint not null auto_increment,
     medico_id bigint not null,
-    paciente_id bigint not null,
+    paciente_id int not null,
     data datetime not null,
 
     primary key(id),
